# openapi/bss-mvp-api-v1.yaml
openapi: 3.1.0
info:
  title: BSS MVP API
  version: 1.1.0
  summary: Implemented Backend MVP contract for the frozen BSS Frontend v1.0.0.
  description: |
    Implemented versioned contract for the functional Backend MVP; deployment remains a separate
    operational decision. The server derives organization,
    role and department scope from the verified session/device identity. Client-supplied
    scope is never authorization evidence. CSV and XLSX are the currently approved export
    formats. PDF is approved as a professional business export in v1.1.
  x-bss-status: MVP_IMPLEMENTED
  x-bss-contract-reviewed-at: 2026-07-17
  x-bss-frontend-release: frontend-v1.0.0
  x-bss-baseline-commit: 91323c7cdbbbbf7b965c4926c94a11af6d31bf62
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: Same-origin application API
...
security:
  - sessionCookie: []
paths:
...
  /dashboard-summary:
    get:
      tags: [Dashboard]
      operationId: getDashboardSummary
      summary: Get Dashboard Summary
      x-bss-roles: [admin, manager, worker, accountant]
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: At most four role-scoped KPI values; every KPI includes a deterministic drill-down target.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'
  /organization:
    get:
      tags: [Organization]
      operationId: getOrganization
      summary: Get Organization
      x-bss-roles: [admin]
      responses:
        '200':
          description: Current organization settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      tags: [Organization]
      operationId: updateOrganization
      summary: Update Organization
      x-bss-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationPatch'
      responses:
        '200':
          description: Updated organization.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '409':
          $ref: '#/components/responses/Conflict'
  /departments:
    get:
      tags: [Organization]
      operationId: listDepartments
      summary: List Departments
      x-bss-roles: [admin, manager, worker]
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Departments inside the effective scope.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Department'
    post:
      tags: [Organization]
      operationId: createDepartment
      summary: Create Department
      x-bss-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentWrite'
      responses:
        '201':
          description: Department created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /departments/{departmentId}:
    patch:
      tags: [Organization]
      operationId: updateDepartment
      summary: Update Department
      x-bss-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/DepartmentIdPath'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentPatch'
      responses:
        '200':
          description: Department updated or blocked while historical relations remain retained.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '409':
          $ref: '#/components/responses/Conflict'
  /holidays:
    get:
      tags: [Organization]
      operationId: listHolidays
      summary: List Holidays
      x-bss-roles: [admin, manager, worker, accountant]
      parameters:
        - $ref: '#/components/parameters/Year'
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Organization holidays used for working-day calculations.
          headers:
            ETag:
              description: Aggregate revision of the requested organization/year calendar.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Holiday'
    put:
      tags: [Organization]
      operationId: replaceHolidaysForYear
      summary: Replace Holidays For Year
      x-bss-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/Year'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              maxItems: 366
              items:
                $ref: '#/components/schemas/HolidayWrite'
      responses:
        '200':
          description: Stored holiday calendar.
          headers:
            ETag:
              description: New aggregate revision of the organization/year calendar.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Holiday'
        '409':
          $ref: '#/components/responses/Conflict'
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List Users
      x-bss-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Organization user accounts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
    post:
      tags: [Users]
      operationId: inviteUser
      summary: Invite User
      x-bss-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInvite'
      responses:
        '202':
          description: Invitation created. A still-unactivated placeholder is safely reused, prior pending invitations are revoked, and the new one-time fragment URL is returned only in this no-store administrator response for secure manual delivery.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInvitation'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /users/{userId}:
    patch:
      tags: [Users]
      operationId: updateUser
      summary: Update User
      x-bss-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPatch'
      responses:
        '200':
          description: Updated user and effective scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          $ref: '#/components/responses/Conflict'
  /workers:
    get:
      tags: [Workers]
      operationId: listWorkers
      summary: List Workers
      x-bss-roles: [admin, manager]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/DepartmentId'
        - $ref: '#/components/parameters/Status'
        - name: search
          in: query
          schema:
            type: string
            maxLength: 100
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Workers inside the verified effective scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerPage'
    post:
      tags: [Workers]
      operationId: createWorker
      summary: Create Worker
      x-bss-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerWrite'
      responses:
        '201':
          description: Worker created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /workers/{workerId}:
    get:
      tags: [Workers]
      operationId: getWorker
      summary: Get Worker
      x-bss-roles: [admin, manager, worker]
      parameters:
        - $ref: '#/components/parameters/WorkerId'
      responses:
        '200':
          description: Worker in effective scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Workers]
      operationId: updateWorker
      summary: Update Worker
      x-bss-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/WorkerId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerWrite'
      responses:
        '200':
          description: Worker updated without deleting history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
        '409':
          $ref: '#/components/responses/Conflict'
  /workers/{workerId}/deactivate:
    post:
      tags: [Workers]
      operationId: deactivateWorker
      summary: Deactivate Worker
      x-bss-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/WorkerId'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Worker blocked while historical records are retained.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
  /shifts:
    get:
      tags: [Shifts]
      operationId: listShifts
      summary: List Shifts
      x-bss-roles: [admin, manager, worker]
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Organization shifts with assignment counts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shift'
    post:
      tags: [Shifts]
      operationId: createShift
      summary: Create Shift
      x-bss-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftWrite'
      responses:
        '201':
          description: Shift created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /shifts/{shiftId}:
    patch:
      tags: [Shifts]
      operationId: updateShift
      summary: Update Shift
      x-bss-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/ShiftId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftWrite'
      responses:
        '200':
          description: Shift updated; historical attendance remains linked to its recorded plan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '409':
          $ref: '#/components/responses/Conflict'
  /workers/{workerId}/activate:
    post:
      tags: [Workers]
      operationId: activateWorker
      summary: Reactivate Worker
      x-bss-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/WorkerId'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Worker reactivated while retained history remains unchanged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
        '409':
          $ref: '#/components/responses/Conflict'
  /rfid-cards/{cardId}/block:
    post:
      tags: [Workers]
      operationId: blockRfidCard
      summary: Block Rfid Card
      x-bss-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/CardId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Card blocked; historical assignments remain available.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfidCard'
...
  /leave-balances:
    get:
      tags: [Leave]
      operationId: listLeaveBalances
      summary: List Leave Balances
      x-bss-roles: [admin, manager, worker, accountant]
      parameters:
        - $ref: '#/components/parameters/Year'
        - $ref: '#/components/parameters/DepartmentId'
        - $ref: '#/components/parameters/WorkerIdQuery'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Server-authoritative annual allowance, approved, planned and available working days in effective scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveBalancePage'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bss_session
      description: Secure, HttpOnly, SameSite session cookie.
    deviceSignature:
      type: apiKey
      in: header
      name: X-BSS-Signature
      description: |
        HMAC-SHA256 v1. The terminal signs UTF-8 bytes of
        METHOD + "\n" + PATH + "\n" + SHA256_HEX(BODY) + "\n" + X-BSS-Timestamp + "\n" + X-BSS-Nonce.
        The signature is lowercase hexadecimal. Server resolves X-BSS-Device-Id inside the tenant-owned
        encrypted credential store, permits at most 300 seconds clock skew and atomically consumes each nonce.
...
  responses:
    Unauthenticated:
      description: Missing, invalid or expired identity.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Authenticated identity is outside the required role or data scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...